Minor/requirements.txt
# Streamlit
streamlit==1.38.0
streamlit-option-menu==0.3.13
streamlit-extras==0.4.3

# LangChain & AI
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface==0.1.0
sentence-transformers==2.6.0
chromadb==0.5.0
huggingface-hub==0.24.0

# PDF Processing
PyPDF2==3.0.1
pypdf==3.17.0
pdfplumber==0.10.3

# Database & Environment
python-dotenv==1.0.0

# Authentication & Security
bcrypt==4.1.1

# Utilities
pydantic>=2.7.4
requests==2.31.0
pandas==2.1.4
plotly==5.18.0